+++ docs/fr_FR/Changelog.rst
@@ -3,6 +3,7 @@ ChangeLog
 
 - Interne: Parser: Correction regression inclusion.
 - Interne: Plusieurs correctifs 'deviceUpdates'.
+- Page avancé: Ajout 'siren level' pour cmde 'Start Warning' (cluster 0502).
 
 230314-BETA-1
 -------------